ci: set up a release pipeline on changes to version.go (#8)
On changes to version.go (in any branch), or when `release:build` is
manually triggered, a release pipeline will execute. The output of this
pipeline is an image in registry.ddbuild.io/commit-headless-prerelease.

Additionally, when a change to version.go lands on main, the prerelease
is promoted to a full release as a tag on
registry.ddbuild.io/commit-headless:v<VERSION>

Each image, regardless of full or pre release status, will contain the
commit-headless binary at `/bin/commit-headless`, and can be used in a
Dockerfile via a COPY --from directive.