feat: build and publish a base CI image during release pipeline
This patch introduces `.ci/publish-image.sh` which is called in the
release pipeline after the image itself is published.

It uses `crane append` to add a new filesystem layer to the
`dd-octo-sts-ci-base` image, adding the architecture specific build of
`commit-headless`.

This results in two images being published on merge:

- registry.ddbuild.io/commit-headless:v$VERSION
- registry.ddbuild.io/commit-headless-ci-image:v$VERSION

The former is designed to be used in pipelines that already have their
own CI image, they can add a `COPY --from` directive to grab the
commit-headless binary. The latter is for pipelines that just need a
simple usable image with dd-octo-sts, commit-headless, and a few other
common tools (such as git, jq, curl).