feat: swap to using the git CLI directly (#14)
This patch replaces `git-go` with calls directly to the `git` CLI. While
it's not ideal to depend on a subprocess, the fact is that using
`git-go` required a lot of lines of code for simple operations, and
often disagreed with `git` on a local repository (which wasn't
reproducible outside of a CI environment).

The `git` CLI commands in-use in this patch are all "plumbing" (versus
"porcelain") commands which generally have stable output: they are
designed explicitly for machine consumption.