ci: set up an action release pipeline (#17)
This patch sets up an action release pipeline on merges to main which
resembles the gitlab release pipeline.

On a merge to main, the workflow kicks in and (assuming everything
builds and tests well), updates the `action` branch with the contents of
`action-template/`.

The workflow leverages the bundled action, which ensures that, at the
very least, we cannot produce a tagged release of the action that cannot
at least release itself.

Note that the action branch contains the binaries built earlier in the
workflow to avoid having to run a composite action.