feat: add --branch-from to create the remote branch if it doesn't exist (#18)
Several cases have already popped up of users creating commits on brand
new branches. I was hoping to avoid bringing this in, but the
implementation is pretty simple.

Both `commit` and `push` gain a new `--branch-from` flag to specify a
commit hash that `--branch` should branch from, if the branch doesn't
already exist on the remote.