feat!: add --head-sha to better control force pushes
To avoid potentially overwriting changes on the remote, allow callers to
specify `--head-sha` (defaulting to the current HEAD on the remote,
which is the previous behavior). This mitigates issues with potentially
overwriting remote changes on long running jobs or active branches,
particularly with jobs that rewrite file contents (such as lint fixes /
auto formatting).

- Replace --branch-from=<sha> with the pair of --head-sha=<sha> and
  --create-branch
- Skip fetching current remote HEAD if --head-sha is supplied